server/requirements.txt
beautifulsoup4==4.8.0
dj-static==0.0.6
Django==2.2.5
django-bootstrap4==1.0.1
django-filter==2.2.0
django-tables2==2.1.0
djangorestframework==3.10.3
mysqlclient==1.4.4
pytz==2019.2
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
static3==0.7.0
unicodecsv==0.14.1
